move objToYAML func to helper.go & log err
